+++ packages/core/src/header-builder/header-util.ts
@@ -30,6 +30,25 @@ export function getHeader(
 }
 
 /**
+ * Find headers in a given header object, if available, idepdendent of the case (lower / upper).
+ * @param keys - Name of the header to be found.
+ * @param headers - Header object to be searched for given key.
+ * @returns - An object containing the given keys (and values) in its original case, as found in `headers` or an empty object if not found.
+ */
+export function getHeaders(
+  keys: string[],
+  headers: Record<string, any> = {}
+): Record<string, any> {
+  return keys.reduce(
+    (filteredHeaders, key) => ({
+      ...filteredHeaders,
+      ...getHeader(key, headers)
+    }),
+    {}
+  );
+}
+
+/**
  * Get the value of a header based on the given key, independent of the case (lower / upper).
  * @param key - Name of the header to be found.
  * @param headers - Header object to be searched for given key.
@@ -72,4 +91,20 @@ export function replaceDuplicateKeys(
         : { [key]: value }
     )
     .reduce((replaced, header) => ({ ...replaced, ...header }), {});
+}
+
+/**
+ * Create a header object by merging two header objects, where the custom headers take precedence. .
+ * @param headers - A base header object that contains the headers that will be compared with `customHeaders`.
+ * @param customHeaders - A header object to be compared with headers. Only headers present in `headers` will be compared.
+ * @returns - An object containing all keys from both the header objects, where headers present in the `customHeaders` are replaced. Note that the case (upper / lower) used by `customHeaders` will be used.
+ */
+export function mergeHeaders(
+  headers: Record<string, any> = {},
+  customHeaders: Record<string, any> = {}
+): Record<string, any> {
+  return {
+    ...replaceDuplicateKeys(headers, customHeaders),
+    ...customHeaders
+  };
 }