+++ CHANGELOG.md
@@ -13,6 +13,30 @@
 -
 
 ## Compatibility Notes
+
+- Deprecate setting the content type header in the constructor of `ODataRequest`s.
+
+## New Functionality
+
+-
+
+## Improvements
+
+- Allow setting default headers in `ODataRequest`s instead of the content type header only.
+
+## Fixed Issues
+
+-
+
+
+# 1.29.0
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/1.29.0<br>
+Blog: TBD<br>
+
+## Compatibility Notes
+
 - [Generator] The interface `EnityTypeForceMandatory` (e.g., `BusinessPartnerTypeForceMandatory`) is removed.
 
 ## New Functionality
@@ -30,7 +54,6 @@
 - [OData] Fix parsing the Etag returned from update requests.
 - [OData] Fix the entity builder so that it can set `undefined`/`null` for optional properties.
 - [OData] Fix wrong service paths for batch requests.
-
 
 # 1.28.2